requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
huepy==1.2.1
Jinja2==3.1.4
livereload==2.7.0
markdown2==2.5.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
node==1.2.2
npm==0.1.1
numpy==2.0.1
odict==1.9.0
opencv-python==4.10.0.84
optional-django==0.1.0
pillow==10.4.0
plumber==1.7
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.1
vite==1.5.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.4.post2
zope.lifecycleevent==5.0
zope.proxy==5.2